+++ packages/react/src/contexts/AccountsProvider.tsx
@@ -7,7 +7,7 @@ import React, {
 } from "react";
 
 import { getMyAccounts } from "@/api/accounts";
-import { STORAGE_KEY_DEFAULT } from "@/constants";
+import { ACCOUNT_HEADER_NAME } from "@/constants";
 import { Account } from "@/types/account";
 import { SaasConfig } from "@/types/config";
 
@@ -20,11 +20,13 @@ export interface AccountsContextType {
   error: boolean;
   accountLoading: boolean;
   meta: {
-    subdomain: string;
+    isAdminApp: boolean;
     isMainApp: boolean;
     mainAppSubdomain: string;
     rootDomain: string;
+    subdomain: string;
   };
+  refetchAccounts: () => void;
   switchAccount: (
     account: Account | null,
     options?: { clearState?: boolean },
@@ -55,11 +57,10 @@ const AccountsProvider = ({ config, userId, children }: Properties) => {
     accounts: accountsConfig,
   } = config;
 
-  const {
-    autoSelectAccount = true,
-    accountStorageKey = STORAGE_KEY_DEFAULT,
-    allowMultipleSessions = true,
-  } = accountsConfig || {};
+  const accountStorageKey = ACCOUNT_HEADER_NAME;
+
+  const { autoSelectAccount = true, allowMultipleSessions = true } =
+    accountsConfig || {};
 
   const subdomain = window.location.hostname.split(".")[0];
   const { isMainApp, isAdminApp } = useMemo(() => {
@@ -170,21 +171,25 @@ const AccountsProvider = ({ config, userId, children }: Properties) => {
     [setLoading, switchAccount, computeNewActiveAccount],
   );
 
+  const fetchMyAccounts = useCallback(() => {
+    setLoading(true);
+
+    getMyAccounts({ apiBaseUrl })
+      .then((accounts) => {
+        updateAccounts(accounts);
+      })
+      .catch((error) => {
+        setError(true);
+      })
+      .finally(() => {
+        setLoading(false);
+      });
+  }, [userId]);
+
   useEffect(() => {
     // ignore account discovery for admin app
     if (!isAdminApp && userId) {
-      setLoading(true);
-
-      getMyAccounts({ apiBaseUrl })
-        .then((accounts) => {
-          updateAccounts(accounts);
-        })
-        .catch((error) => {
-          setError(true);
-        })
-        .finally(() => {
-          setLoading(false);
-        });
+      fetchMyAccounts();
     }
   }, [userId]);
 
@@ -197,11 +202,13 @@ const AccountsProvider = ({ config, userId, children }: Properties) => {
         error,
         accountLoading,
         meta: {
-          subdomain,
+          isAdminApp,
           isMainApp,
           mainAppSubdomain,
           rootDomain,
+          subdomain,
         },
+        refetchAccounts: fetchMyAccounts,
         switchAccount,
         updateAccounts,
       }}