+++ packages/react/src/views/AcceptInvitation/index.ts
@@ -1,1 +1,3 @@
 export * from "./AcceptInvitation";
+export * from "./JoinInvitation";
+export * from "./SignupInvitation";